[ecore][cpp] define a EObjectList interface
