+++ ecore/lib/CMakeFiles.txt
@@ -26,6 +26,7 @@ set(HEADER_ECORE_FILES
     src/ecore/ENotificationChain.hpp
     src/ecore/ENotifier.hpp
     src/ecore/ENotifyingList.hpp
+    src/ecore/EObjectList.hpp
     src/ecore/EPackageRegistry.hpp
     src/ecore/EPackageResourceRegistry.hpp
     src/ecore/EResource.hpp
